Chem: Move code parts into chem-atom-picker-utils.ts
