+++ packages/Chem/src/rendering/rdkit-cell-renderer.ts
@@ -32,6 +32,9 @@ import {IMolContext, getMolSafe} from '../utils/mol-creation_rdkit';
 import {getGridCellColTemp} from '../utils/ui-utils';
 import {errInfo} from '../utils/err-info';
 import {mapAtomIndices2Dto3D} from '../utils/atom-index-mapper';
+import {
+  findNearestAtom, computeSelectedBonds, extractAtomPositionsFromSvg,
+} from '../utils/chem-atom-picker-utils';
 
 import {_package} from '../package';
 
@@ -782,7 +785,7 @@ M  END
     const cellInfo = this._getCellAtomPositions(molString, bounds.width, bounds.height);
     if (!cellInfo || cellInfo.positions.size === 0) return null;
 
-    const nearest = this._findNearestAtom(cellInfo.positions, pointerCellX, pointerCellY);
+    const nearest = findNearestAtom(cellInfo.positions, pointerCellX, pointerCellY);
     if (nearest === null) return null;
 
     return {grid, gridCell, cellInfo, nearest, pointerCellX, pointerCellY};
@@ -905,25 +908,6 @@ M  END
   /** Returns the index of the atom whose center is nearest to the click
    *  point, but only if it falls within `CLICK_TOLERANCE_PX` (CSS pixels).
    *  Returns null if the user clicked too far from any atom. */
-  private _findNearestAtom(
-    positions: Map<number, {x: number, y: number}>,
-    clickX: number, clickY: number,
-  ): number | null {
-    const CLICK_TOLERANCE_PX = 12;
-    const tolSq = CLICK_TOLERANCE_PX * CLICK_TOLERANCE_PX;
-    let nearestIdx: number | null = null;
-    let nearestDistSq = tolSq;
-    for (const [idx, p] of positions.entries()) {
-      const dx = p.x - clickX;
-      const dy = p.y - clickY;
-      const dsq = dx * dx + dy * dy;
-      if (dsq < nearestDistSq) {
-        nearestDistSq = dsq;
-        nearestIdx = idx;
-      }
-    }
-    return nearestIdx;
-  }
 
   /** Returns the persistent (Alt+painted) atoms for a given row. These
    *  are the atoms stored in the provider's `__atoms` set. */
@@ -967,7 +951,7 @@ M  END
       const highlightAtomColors: {[k: number]: number[]} = {};
       for (const a of atomsArr) highlightAtomColors[a] = pickColor;
       const {bondsArr, highlightBondColors} =
-        RDKitCellRenderer._computeSelectedBonds(prov.__atoms, bondAtoms, pickColor);
+        computeSelectedBonds(prov.__atoms, bondAtoms, pickColor);
 
       // Replace the getSubstruct callback with updated atoms.
       prov.getSubstruct = (ridx) => {
@@ -1087,7 +1071,7 @@ M  END
     const highlightAtomColors: {[k: number]: number[]} = {};
     for (const a of atomsArr) highlightAtomColors[a] = pickColor;
     const {bondsArr, highlightBondColors} =
-      RDKitCellRenderer._computeSelectedBonds(current, bondAtoms, pickColor);
+      computeSelectedBonds(current, bondAtoms, pickColor);
     const provider: TaggedProvider = {
       getSubstruct: (ridx) => {
         if (!this._isPickerActive(col)) return undefined;
@@ -1134,21 +1118,6 @@ M  END
    *  bond indices where BOTH endpoint atoms are in the selection, and a
    *  matching highlightBondColors map with every selected bond painted the
    *  same color. Shared by _updateRowSelection and _toggleAtomInRow. */
-  private static _computeSelectedBonds(
-    atoms: Set<number>,
-    bondAtoms: Map<number, [number, number]>,
-    color: number[],
-  ): {bondsArr: number[], highlightBondColors: {[k: number]: number[]}} {
-    const bondsArr: number[] = [];
-    const highlightBondColors: {[k: number]: number[]} = {};
-    for (const [bondIdx, [a1, a2]] of bondAtoms.entries()) {
-      if (atoms.has(a1) && atoms.has(a2)) {
-        bondsArr.push(bondIdx);
-        highlightBondColors[bondIdx] = color;
-      }
-    }
-    return {bondsArr, highlightBondColors};
-  }
 
   /**
    * Merges the freshly boxed atoms into the row's current atom-picker selection
@@ -1189,7 +1158,7 @@ M  END
     // selected region read as a connected sub-molecule rather than a
     // scatter of disconnected atoms.
     const {bondsArr, highlightBondColors} =
-      RDKitCellRenderer._computeSelectedBonds(current, bondAtoms, pickColor);
+      computeSelectedBonds(current, bondAtoms, pickColor);
 
     const provider: TaggedProvider = {
       getSubstruct: (ridx) => {
@@ -1302,7 +1271,7 @@ M  END
       try {
         const svgEl = host.querySelector('svg') as SVGSVGElement | null;
         if (!svgEl) return null;
-        const extracted = this._extractAtomPositionsFromSvg(svgEl);
+        const extracted = extractAtomPositionsFromSvg(svgEl);
         // Convert from DPR-scaled SVG coordinates back to cell-local CSS
         // pixels so they can be compared against mouse event coordinates.
         const positions = new Map<number, {x: number, y: number}>();
@@ -1318,76 +1287,6 @@ M  END
     } catch {
       return null;
     }
-  }
-
-  private _extractAtomPositionsFromSvg(svgEl: SVGSVGElement):
-      {positions: Map<number, {x: number, y: number}>, bondAtoms: Map<number, [number, number]>} {
-    const positions = new Map<number, {x: number, y: number}>();
-    /** bond index → the two atom indices this bond connects. Populated
-     *  below from `class="bond-K atom-A atom-B"` on RDKit's bond paths. */
-    const bondAtoms = new Map<number, [number, number]>();
-
-    // Text-labeled heteroatoms
-    const texts = svgEl.querySelectorAll('text[class*="atom-"]');
-    for (let i = 0; i < texts.length; i++) {
-      const t = texts[i];
-      const cls = t.getAttribute('class') || '';
-      const m = /(?:^|\s)atom-(\d+)/.exec(cls);
-      if (!m) continue;
-      const idx = parseInt(m[1], 10);
-      try {
-        const bb = (t as unknown as SVGGraphicsElement).getBBox();
-        positions.set(idx, {x: bb.x + bb.width / 2, y: bb.y + bb.height / 2});
-      } catch {/* ignore */}
-    }
-
-    // Carbons — average bond endpoints extracted from <path> d attributes.
-    // Also populate bondAtoms as a side-effect so we don't need to re-walk
-    // the path list when the picker wants bond connectivity.
-    const bondEnds = new Map<number, Array<{x: number, y: number}>>();
-    const bondPaths = svgEl.querySelectorAll('path[class*="bond-"]');
-    for (let i = 0; i < bondPaths.length; i++) {
-      const p = bondPaths[i];
-      const cls = p.getAttribute('class') || '';
-      const atomIds: number[] = [];
-      const atomMatches = cls.match(/atom-(\d+)/g) || [];
-      for (const mm of atomMatches) {
-        const n = parseInt(mm.slice(5), 10);
-        if (!Number.isNaN(n)) atomIds.push(n);
-      }
-      if (atomIds.length !== 2) continue;
-
-      // Record the bond-index → atom-pair mapping. RDKit emits multiple
-      // path segments per bond for aromatic/double/triple bonds (one
-      // segment per line), all sharing the same bond-K class — but
-      // `bondAtoms.set` is idempotent for the same key, so overwriting is
-      // harmless.
-      const bondMatch = /(?:^|\s)bond-(\d+)/.exec(cls);
-      if (bondMatch) {
-        const bondIdx = parseInt(bondMatch[1], 10);
-        bondAtoms.set(bondIdx, [atomIds[0], atomIds[1]]);
-      }
-
-      const d = p.getAttribute('d') || '';
-      const coords: Array<{x: number, y: number}> = [];
-      const re = /[ML]\s*([\-\d.]+)[\s,]+([\-\d.]+)/g;
-      let mm: RegExpExecArray | null;
-      while ((mm = re.exec(d)) !== null)
-        coords.push({x: parseFloat(mm[1]), y: parseFloat(mm[2])});
-      if (coords.length < 2) continue;
-      if (!bondEnds.has(atomIds[0])) bondEnds.set(atomIds[0], []);
-      if (!bondEnds.has(atomIds[1])) bondEnds.set(atomIds[1], []);
-      bondEnds.get(atomIds[0])!.push(coords[0]);
-      bondEnds.get(atomIds[1])!.push(coords[coords.length - 1]);
-    }
-    for (const [idx, pts] of bondEnds.entries()) {
-      if (positions.has(idx)) continue;
-      const cx = pts.reduce((s, p) => s + p.x, 0) / pts.length;
-      const cy = pts.reduce((s, p) => s + p.y, 0) / pts.length;
-      positions.set(idx, {x: cx, y: cy});
-    }
-
-    return {positions, bondAtoms};
   }
 }
